hotshot_testing
Module block_builder

Test implementation of block builder

Re-exports

pub use random::RandomBuilderImplementation;

pub use simple::SimpleBuilderImplementation;

Modules

random

simple

Structs

BlockEntry 🔒

Entry for a built block

Traits

BuilderTask

TestBuilderImplementation

Functions

build_block 🔒

Helper function to construct all builder data structures from a list of transactions

run_builder_source

Construct a tide disco app that mocks the builder API 0.1 + 0.3.

run_builder_source_0_1

Construct a tide disco app that mocks the builder API 0.1.