Function hotshot_testing::block_builder::run_builder_source

pub fn run_builder_source<TYPES, Source>(
    url: Url,
    change_receiver: Receiver<BuilderChange>,
    source: Source,
)
where
    TYPES: NodeType,
    <TYPES as NodeType>::InstanceState: Default,
    Source: Clone + Send + Sync + ReadState + 'static,
    <Source as ReadState>::State: Sync + Send + BuilderDataSource<TYPES> + BuilderDataSource<TYPES>,

Construct a tide disco app that mocks the builder API 0.1 + 0.3.

Panics

If constructing and launching the builder fails for any reason