Module hotshot_testing::consistency_task

task that checks leaves received across all nodes from decide events for consistency

Structs

ConsistencyTask

Data availability task state

Functions

invert_network_map 🔒

sanitize_network_map 🔒

Validate that each node has only produced one unique leaf per view, and produce a NetworkMapSanitized.

sanitize_node_map 🔒

Validate that the NodeMap only has a single leaf per view.

sanitize_view_map 🔒

validate_node_map 🔒

For a NodeMapSanitized, we validate that each leaf extends the preceding leaf.

Type Aliases

NetworkMap

A map from node ids to NodeMaps; note that the latter may have multiple leaves per view in principle.

NetworkMapSanitized

A map from node ids to NodeMapSanitizeds; the latter has been sanitized validated to have a single leaf per view.

NodeMap

Map from views to leaves for a single node, allowing multiple leaves for each view (because the node may a priori send us multiple leaves for a given view).

NodeMapSanitized

A sanitized map from views to leaves for a single node, with only a single leaf per view.

ViewMap

ViewMapSanitized

A view map, sanitized to have exactly one leaf per view.