Module hotshot_testing::overall_safety_task

task that’s consuming events and asserting safety

Structs

InconsistentBlocksSnafu 🔒

SNAFU context selector for the OverallSafetyTaskErr::InconsistentBlocks variant

InconsistentStatesSnafu 🔒

SNAFU context selector for the OverallSafetyTaskErr::InconsistentStates variant

InconsistentTxnsNumSnafu 🔒

SNAFU context selector for the OverallSafetyTaskErr::InconsistentTxnsNum variant

MismatchedLeafSnafu 🔒

SNAFU context selector for the OverallSafetyTaskErr::MismatchedLeaf variant

NotEnoughDecidesSnafu 🔒

SNAFU context selector for the OverallSafetyTaskErr::NotEnoughDecides variant

OverallSafetyPropertiesDescription

cross node safety properties

OverallSafetyTask

Data availability task state

RoundCtx

context for a round TODO eventually we want these to just be futures that we poll when things are event driven this context will be passed around

RoundResult

Result of running a round of consensus

TooManyFailuresSnafu 🔒

SNAFU context selector for the OverallSafetyTaskErr::TooManyFailures variant

Enums

OverallSafetyTaskErr

possible errors

ViewStatus

the status of a view

Type Aliases

StateAndBlock

convenience type alias for state and block