Type Alias hotshot_testing::predicates::ConsensusTaskTestState

type ConsensusTaskTestState = ConsensusTaskState<TestTypes, MemoryImpl, SystemContextHandle<TestTypes, MemoryImpl>>;

Aliased Type

struct ConsensusTaskTestState {

    pub public_key: VerKey,
    pub private_key: SignKey,
    pub consensus: Arc<RwLock<Consensus<TestTypes>>>,
    pub timeout: u64,
    pub cur_view: ViewNumber,
    pub payload_commitment_and_metadata: Option<CommitmentAndMetadata<TestBlockPayload>>,
    pub quorum_network: Arc<MemoryNetwork<Message<TestTypes>, VerKey>>,
    pub committee_network: Arc<MemoryNetwork<Message<TestTypes>, VerKey>>,
    pub timeout_membership: Arc<GeneralStaticCommittee<TestTypes, VerKey>>,
    pub quorum_membership: Arc<GeneralStaticCommittee<TestTypes, VerKey>>,
    pub committee_membership: Arc<GeneralStaticCommittee<TestTypes, VerKey>>,
    pub api: SystemContextHandle<TestTypes, MemoryImpl>,
    pub _pd: PhantomData<MemoryImpl>,
    pub vote_collector: RwLock<Option<VoteCollectionTaskState<TestTypes, SimpleVote<TestTypes, QuorumData<TestTypes>>, SimpleCertificate<TestTypes, QuorumData<TestTypes>, SuccessThreshold>>>>,
    pub timeout_vote_collector: RwLock<Option<VoteCollectionTaskState<TestTypes, SimpleVote<TestTypes, TimeoutData<TestTypes>>, SimpleCertificate<TestTypes, TimeoutData<TestTypes>, SuccessThreshold>>>>,
    pub timeout_task: Option<JoinHandle<()>>,
    pub timeout_cert: Option<SimpleCertificate<TestTypes, TimeoutData<TestTypes>, SuccessThreshold>>,
    pub upgrade_cert: Option<SimpleCertificate<TestTypes, UpgradeProposalData<TestTypes>, UpgradeThreshold>>,
    pub view_sync_cert: Option<SimpleCertificate<TestTypes, ViewSyncFinalizeData<TestTypes>, SuccessThreshold>>,
    pub decided_upgrade_cert: Option<SimpleCertificate<TestTypes, UpgradeProposalData<TestTypes>, UpgradeThreshold>>,
    pub current_network_version: Version,
    pub output_event_stream: Sender<Event<TestTypes>>,
    pub current_proposal: Option<QuorumProposal<TestTypes>>,
    pub id: u64,
    pub storage: Arc<RwLock<TestStorage<TestTypes>>>,

}

Fields
public_key: VerKey

Our public key

private_key: SignKey

Our Private Key

consensus: Arc<RwLock<Consensus<TestTypes>>>

Reference to consensus. The replica will require a write lock on this.

timeout: u64

View timeout from config.

cur_view: ViewNumber

View number this view is executing in.

payload_commitment_and_metadata: Option<CommitmentAndMetadata<TestBlockPayload>>

The commitment to the current block payload and its metadata submitted to DA.

quorum_network: Arc<MemoryNetwork<Message<TestTypes>, VerKey>>

Network for all nodes

committee_network: Arc<MemoryNetwork<Message<TestTypes>, VerKey>>

Network for DA committee

timeout_membership: Arc<GeneralStaticCommittee<TestTypes, VerKey>>

Membership for Timeout votes/certs

quorum_membership: Arc<GeneralStaticCommittee<TestTypes, VerKey>>

Membership for Quorum Certs/votes

committee_membership: Arc<GeneralStaticCommittee<TestTypes, VerKey>>

Membership for DA committee Votes/certs

api: SystemContextHandle<TestTypes, MemoryImpl>

Consensus api

_pd: PhantomData<MemoryImpl>

needed to typecheck

vote_collector: RwLock<Option<VoteCollectionTaskState<TestTypes, SimpleVote<TestTypes, QuorumData<TestTypes>>, SimpleCertificate<TestTypes, QuorumData<TestTypes>, SuccessThreshold>>>>

Current Vote collection task, with it’s view.

timeout_vote_collector: RwLock<Option<VoteCollectionTaskState<TestTypes, SimpleVote<TestTypes, TimeoutData<TestTypes>>, SimpleCertificate<TestTypes, TimeoutData<TestTypes>, SuccessThreshold>>>>

Current timeout vote collection task with its view

timeout_task: Option<JoinHandle<()>>

timeout task handle

timeout_cert: Option<SimpleCertificate<TestTypes, TimeoutData<TestTypes>, SuccessThreshold>>

last Timeout Certificate this node formed

upgrade_cert: Option<SimpleCertificate<TestTypes, UpgradeProposalData<TestTypes>, UpgradeThreshold>>

last Upgrade Certificate this node formed

view_sync_cert: Option<SimpleCertificate<TestTypes, ViewSyncFinalizeData<TestTypes>, SuccessThreshold>>

last View Sync Certificate this node formed

decided_upgrade_cert: Option<SimpleCertificate<TestTypes, UpgradeProposalData<TestTypes>, UpgradeThreshold>>

most recent decided upgrade certificate

current_network_version: Version

The current version of the network. Updated on view change based on the most recent decided upgrade certificate.

output_event_stream: Sender<Event<TestTypes>>

Output events to application

current_proposal: Option<QuorumProposal<TestTypes>>

The most recent proposal we have, will correspond to the current view if Some() Will be none if the view advanced through timeout/view_sync

id: u64

The node’s id

storage: Arc<RwLock<TestStorage<TestTypes>>>

This node’s storage ref