Type Alias hotshot_testing::predicates::upgrade_with_proposal::QuorumProposalTaskTestState

type QuorumProposalTaskTestState = QuorumProposalTaskState<TestTypes, MemoryImpl, TestVersions>;

Aliased Type

struct QuorumProposalTaskTestState {

    pub latest_proposed_view: ViewNumber,
    pub cur_epoch: Option<EpochNumber>,
    pub proposal_dependencies: BTreeMap<ViewNumber, JoinHandle<()>>,
    pub instance_state: Arc<TestInstanceState>,
    pub membership: Arc<RwLock<StaticCommittee<TestTypes>>>,
    pub public_key: VerKey,
    pub private_key: SignKey,
    pub timeout: u64,
    pub storage: Arc<RwLock<TestStorage<TestTypes>>>,
    pub consensus: OuterConsensus<TestTypes>,
    pub id: u64,
    pub formed_upgrade_certificate: Option<SimpleCertificate<TestTypes, UpgradeProposalData<TestTypes>, UpgradeThreshold>>,
    pub upgrade_lock: UpgradeLock<TestTypes, TestVersions>,
    pub epoch_height: u64,
    pub highest_qc: SimpleCertificate<TestTypes, QuorumData2<TestTypes>, SuccessThreshold>,

}

Fields
latest_proposed_view: ViewNumber

Latest view number that has been proposed for.

cur_epoch: Option<EpochNumber>

Current epoch

proposal_dependencies: BTreeMap<ViewNumber, JoinHandle<()>>

Table for the in-progress proposal dependency tasks.

instance_state: Arc<TestInstanceState>

Immutable instance state

membership: Arc<RwLock<StaticCommittee<TestTypes>>>

Membership for Quorum Certs/votes

public_key: VerKey

Our public key

private_key: SignKey

Our Private Key

timeout: u64

View timeout from config.

storage: Arc<RwLock<TestStorage<TestTypes>>>

This node’s storage ref

consensus: OuterConsensus<TestTypes>

Shared consensus task state

id: u64

The node’s id

formed_upgrade_certificate: Option<SimpleCertificate<TestTypes, UpgradeProposalData<TestTypes>, UpgradeThreshold>>

The most recent upgrade certificate this node formed. Note: this is ONLY for certificates that have been formed internally, so that we can propose with them.

Certificates received from other nodes will get reattached regardless of this fields, since they will be present in the leaf we propose off of.

upgrade_lock: UpgradeLock<TestTypes, TestVersions>

Lock for a decided upgrade

epoch_height: u64

Number of blocks in an epoch, zero means there are no epochs

highest_qc: SimpleCertificate<TestTypes, QuorumData2<TestTypes>, SuccessThreshold>

The highest_qc we’ve seen at the start of this task