Module hotshot_testing::test_description

structs and infra to describe the tests to be written

Structs

DetailedTestDescriptionBuilder

fine-grained spec of test including what should be run every round and how to generate more rounds

GeneralTestDescriptionBuilder

! public infra for describing tests TODO this should be taking in a election config

HotShotSnafu 🔒

SNAFU context selector for the RoundError::HotShot variant

TestDescription

Description of a test. Contains all metadata necessary to execute test

TimingData

data describing how a round should be timed.

Enums

RoundError 🔒

Functions

default_randomized_ids_to_round

generate a randomized set of transactions each round

default_submitter_id_to_round

given a description of rounds, generates such rounds args

get_threshold

given num_nodes, calculate min number of honest nodes for consensus to function properly

get_tolerance

given num_nodes, calculate max number of byzantine nodes

Type Definitions

GenRunner

type alias for generating a TestRunner

TestSetup

type alias for doing setup for a consensus round