hotshot_testing
Module test_runner

runner

Structs

LateNodeContextParameters

This type combines all of the parameters needed to build the context for a node that started late during a unit test or integration test.

LateStartNode

A yet-to-be-started node that participates in tests

Node

a node participating in a test

TestRunner

The runner of a test network spin up and down nodes, execute rounds

Enums

LateNodeContext

The late node context dictates how we’re building a node that started late during the test.

Traits

TaskErr