Crate hotshot_testing

Testing infrastructure for HotShot

Modules

block_builder

Test implementation of block builder

completion_task

task that decides when things are complete

consistency_task

task that checks leaves received across all nodes from decide events for consistency

helpers

Helpers for initializing system context handle and building tasks.

overall_safety_task

task that’s consuming events and asserting safety

predicates

predicates to use in tests

script

scripting harness for tests

spinning_task

task to spin nodes up and down

test_builder

builder

test_launcher

launcher

test_runner

runner

test_task

the TestTask struct and associated trait/functions

txn_task

task that’s submitting transactions to the stream

view_generator

view generator for tests

view_sync_task

task for checking if view sync got activated

Macros

all_predicates

random

serial