Crate hotshot_testing

Testing harness for the hotshot repository

To build a test environment you can create a [TestLauncher] instance. This launcher can be configured to have a custom networking layer, initial state, etc.

Calling TestLauncher::launch() will turn this launcher into a [TestRunner], which can be used to start and stop nodes, send transacstions, etc.

Node that TestLauncher::launch() is only available if the given NETWORK, STATE and STORAGE are correct.

Purpose

Infrastructure and integration tests for hotshot.

Usage

The overall control flow is:

TestBuilder::default().build() ->  TestLauncher::launch()      -> TestRunner::execute()
|                                  |                              |
- easy override setup fn           - more explicit overrides      - executes the test
|                                    | for networking, storage,
- easy override correctness fn       | hooks/overrides etc
|
- easily add in hooks
|
- easily override launching

Easily overriding setup/correctness checks/hooks and launching is all done by anonymous functions. Fairly sane and configurable setup and correct check functions may be generated from the round builder. The intended workflow should look like:

use std::sync::Arc;
use futures::FutureExt;
use hotshot_testing::test_types::StaticNodeImplType;
use hotshot_testing::round::RoundHook;
use hotshot_testing::test_types::StaticCommitteeTestTypes;
use hotshot_testing::test_builder::TestBuilder;
use hotshot_testing::test_builder::TestMetadata;

async {
    // specify general characteristics of the test in TestMetadata
    let metadata = TestMetadata {
        total_nodes: 10,
        start_nodes: 10,
        num_succeeds: 5,
        failure_threshold: 10,
        ..Default::default()
    };

    // construct the builder
    let mut test_builder = TestBuilder {
        metadata,
        /// we could build a check
        check: None,
        /// or a round setup if we want
        setup: None
    };

    // construct the launcher
    // this may be used to manually override any of the round functions
    let test_launcher = test_builder.build::<StaticCommitteeTestTypes, StaticNodeImplType>();

    /// now let's add in a custom hook to print some debugging information at the beginning
    /// of each view
    let hook =
        RoundHook(Arc::new(move |_runner, ctx| {
            async move {
                tracing::error!("Context for this view is {:#?})", ctx);
                Ok(())
            }
            .boxed_local()
        }));

    /// add the hook, launch the test, then run it.
    test_launcher.push_hook(hook).launch().run_test().await.unwrap();

};

See TODO for examples.

Modules

network_reliability

implementations of various networking models

round

describe a round of consensus

round_builder

helper functions to build a round

test_builder

structs and infra to describe the tests to be written

test_errors

errors for tests

test_launcher

test launcher infrastructure

test_runner

test runner

test_types

set of commonly used test types for our tests