Module hotshot_testing::byzantine::byzantine_behaviour

Byzantine definitions and implementations of different behaviours

Structs

BadProposalViewDos

An EventTransformerState that multiplies QuorumProposalSend events, incrementing the view number of the proposal

DishonestDa

An EventHandlerState that modifies view number on the certificate of DacSend event to that of a future view

DishonestLeader

An EventHandlerState that modifies justify_qc on QuorumProposalSend to that of a previous view to mock dishonest leader

DishonestVoter

An EventHandlerState that will send a vote for a bad proposal

DishonestVoting

An EventHandlerState that modifies view number on the vote of QuorumVoteSend event to that of a future view and correctly signs the vote

DoubleProposeVote

An EventHandlerState that doubles the QuorumVoteSend and QuorumProposalSend events

ViewDelay

View delay configuration