Module hotshot_testing::helpers

Helpers for initializing system context handle and building tasks.

Functions

build_assembled_sig

create signature

build_cert

create certificate

build_da_certificate

build_fake_view_with_leaf

This function will create a fake View from a provided [Leaf].

build_fake_view_with_leaf_and_state

This function will create a fake View from a provided [Leaf] and state.

build_payload_commitment

build_system_handle

create the SystemContextHandle from a node id, with no epochs

build_system_handle_from_launcher

create the SystemContextHandle from a node id and TestLauncher

build_vid_proposal

TODO: https://github.com/EspressoSystems/HotShot/issues/2821

da_payload_commitment

key_pair_for_id

get the keypair for a node id

permute_input_with_index_order

This function permutes the provided input vector inputs, given some order provided within the order vector.

vid_payload_commitment

vid_scheme_from_view_number

initialize VID

vid_share