Type Alias hotshot_testing::predicates::upgrade_with_vote::QuorumVoteTaskTestState

type QuorumVoteTaskTestState = QuorumVoteTaskState<TestTypes, MemoryImpl, TestVersions>;

Aliased Type

struct QuorumVoteTaskTestState {

    pub public_key: VerKey,
    pub private_key: SignKey,
    pub consensus: OuterConsensus<TestTypes>,
    pub instance_state: Arc<TestInstanceState>,
    pub latest_voted_view: ViewNumber,
    pub vote_dependencies: BTreeMap<ViewNumber, JoinHandle<()>>,
    pub network: Arc<MemoryNetwork<VerKey>>,
    pub membership: Arc<RwLock<StaticCommittee<TestTypes>>>,
    pub drb_computation: Option<(EpochNumber, JoinHandle<[u8; 32]>)>,
    pub output_event_stream: Sender<Event<TestTypes>>,
    pub id: u64,
    pub consensus_metrics: Arc<ConsensusMetricsValue>,
    pub storage: Arc<RwLock<TestStorage<TestTypes>>>,
    pub upgrade_lock: UpgradeLock<TestTypes, TestVersions>,
    pub epoch_height: u64,

}

Fields
public_key: VerKey

Public key.

private_key: SignKey

Private Key.

consensus: OuterConsensus<TestTypes>

Reference to consensus. The replica will require a write lock on this.

instance_state: Arc<TestInstanceState>

Immutable instance state

latest_voted_view: ViewNumber

Latest view number that has been voted for.

vote_dependencies: BTreeMap<ViewNumber, JoinHandle<()>>

Table for the in-progress dependency tasks.

network: Arc<MemoryNetwork<VerKey>>

The underlying network

membership: Arc<RwLock<StaticCommittee<TestTypes>>>

Membership for Quorum certs/votes and DA committee certs/votes.

drb_computation: Option<(EpochNumber, JoinHandle<[u8; 32]>)>

In-progress DRB computation task.

output_event_stream: Sender<Event<TestTypes>>

Output events to application

id: u64

The node’s id

consensus_metrics: Arc<ConsensusMetricsValue>

The consensus metrics

storage: Arc<RwLock<TestStorage<TestTypes>>>

Reference to the storage.

upgrade_lock: UpgradeLock<TestTypes, TestVersions>

Lock for a decided upgrade

epoch_height: u64

Number of blocks in an epoch, zero means there are no epochs