Module hotshot_testing::test_task

the TestTask struct and associated trait/functions

Structs

TestTask

A basic task which loops waiting for events to come from event_receiver and then handles them using it’s state It sends events to other Tasks through event_sender This should be used as the primary building block for long running or medium running tasks (i.e. anything that can’t be described as a dependency task)

Enums

TestEvent

TestResult

enum describing how the tasks completed

Traits

TestTaskState

Type for mutable task state that can be used as the state for a Task

TestTaskStateSeed

Functions

add_network_message_test_task

Add the network task to handle messages and publish events.

Type Aliases

AnyTestTaskState

Type alias for type-erased TestTaskState to be used for dynamic dispatch