hotshot_testing::test_task
Trait TestTaskState

pub trait TestTaskState: Send {
    type Event: Clone + Send + Sync;
    type Error: Display;

    // Required methods
    fn handle_event<'life0, 'async_trait>(
        &'life0 mut self,
        __arg1: (Self::Event, usize),
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn check<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = TestResult> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Type for mutable task state that can be used as the state for a Task

Required Associated Types

type Event: Clone + Send + Sync

Type of event sent and received by the task

type Error: Display

Type of error produced by the task

Required Methods

fn handle_event<'life0, 'async_trait>( &'life0 mut self, __arg1: (Self::Event, usize), ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Handles an event from one of multiple receivers.

fn check<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = TestResult> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Check the result of the test.